chore(refactor): centralize all providers common method
