+++ src/rpc.ts
@@ -1,13 +1,10 @@
 import express from 'express';
 import Promise from 'bluebird';
+import { capture } from '@snapshot-labs/snapshot-sentry';
 import { MAX, rpcError, rpcSuccess } from './utils';
-import { set as setFleek } from './providers/fleek';
-import { set as setInfura } from './providers/infura';
-import { set as setWeb3Storage } from './providers/web3storage';
-import { set as set4everland } from './providers/4everland';
 import { set as setAws } from './aws';
 import { stats } from './stats';
-import { capture } from '@snapshot-labs/snapshot-sentry';
+import { JSON_PROVIDERS, default as set } from './providers/';
 
 const router = express.Router();
 
@@ -16,12 +13,7 @@ router.post('/', async (req, res) => {
   try {
     const size = Buffer.from(JSON.stringify(params)).length;
     if (size > MAX) return rpcError(res, 400, 'File too large', id);
-    const result = await Promise.any([
-      setFleek(params),
-      setInfura(params),
-      setWeb3Storage(params),
-      set4everland(params)
-    ]);
+    const result = await Promise.any(set(JSON_PROVIDERS, params));
     await setAws(result.cid, params);
     stats.providers[result.provider] = (stats.providers[result.provider] || 0) + 1;
     stats.total += 1;
